requirements.txt
apache-airflow==2.7.1
pandas==2.1.1
psycopg2-binary==2.9.7

dbt-core==1.7.1
dbt-postgres==1.7.1

python-dotenv==1.0.0
